feat(snap): Add snap support for reader discovery
Executing the command edgex-device-rfid-llrp.discover-readers will run a script which
scans for the available subnets and then scans them for available RFID drivers.

This snap requires the network-observe and network-control interfaces to work and
when they have been added a hook runs device discovery automatically.

To test this, do
 snapcraft --use-lxd
 sudo snap install --dangerous *.snap
 sudo snap connect edgex-device-rfid-llrp:network-observe :network-observe
 sudo snap connect edgex-device-rfid-llrp:network-control :network-control

That will run the device discovery and edgex-cli device list should then list
the found devices.

Signed-off-by: Siggi Skulason <siggi.skulason@canonical.com>